// src/types/card.ts
/* eslint-disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

export interface Card {
  id: string;
  dominio: number;
  task_statement: string;
  leccion: number;
  tema: string;
  tipo:
    | "definicion"
    | "comparacion"
    | "servicio_aws"
    | "caso_de_uso"
    | "escenario_examen"
    | "calculo"
    | "por_que"
    | "arquitectura"
    | "secuencia"
    | "trampa_examen";
  nivel: 1 | 2 | 3 | 4 | 5;
  frente: string;
  reverso: Reverso;
  servicios_aws: string[];
  tags: string[];
  relacionadas: string[];
  fuente: string;
}
export interface Reverso {
  respuesta_corta: string;
  explicacion: string;
  glosario_refs: string[];
  puntos_clave: string[];
  descartes: Descarte[];
  tip_examen: string;
}
export interface Descarte {
  opcion: string;
  por_que_no: string;
}