// src/types/glossary.ts
/* eslint-disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

export interface Glossary {
  version: string;
  nota: string;
  entries: GlossaryEntry[];
}
export interface GlossaryEntry {
  id: string;
  termino: string;
  definicion: string;
  aliases: string[];
  /**
   * @minItems 1
   */
  dominios: [number, ...number[]];
}